+++ transportation-service/src/main/java/com/ftcs/transportation/trip_matching/repository/TripMatchingFinalRepository.java
@@ -7,5 +7,5 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-public interface TripMatchingFinalRepository extends JpaRepository<TripMatchingFinal, Integer> {
+public interface TripMatchingFinalRepository extends JpaRepository<TripMatchingFinal, Long> {
 }
